pyproject.toml
[project]
name = "aicodegencrew"
version = "0.3.0"
description = "AI-powered SDLC automation using CrewAI with local Ollama models"
authors = [{ name = "Aymen Mastouri", email = "aymen.mastouri@icloud.com" }]
requires-python = ">=3.10,<3.13"
readme = "README.md"
license = { text = "Proprietary" }
keywords = ["ai", "crewai", "ollama", "sdlc", "code-generation", "rag"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: Other/Proprietary License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Software Development :: Documentation",
]

dependencies = [
    # Core AI/ML
    "crewai[tools]>=1.9.0",
    "ollama>=0.4.5",
    # Vector DB & Embeddings
    "chromadb>=1.1.0",
    # MCP Server
    "mcp[cli]>=1.2.0",
    # Configuration & Utilities
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "jinja2>=3.1.2",
    "requests>=2.31.0",
    "pyyaml>=6.0",
    # Document format conversion (MD → Confluence/AsciiDoc/HTML)
    "mistune>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
parsers = [
    "python-docx>=1.0.0",  # DOCX parsing
    "openpyxl>=3.1.0",      # Excel parsing
]

[project.scripts]
aicodegencrew = "aicodegencrew.main:run"
aicodegencrew-mcp = "aicodegencrew.mcp.server:run_server"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aicodegencrew"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "phase0: Phase 0 tests (indexing)",
    "phase1: Phase 1 tests (facts extraction)",
    "phase2: Phase 2 tests (architecture analysis)",
    "phase3: Phase 3 tests (architecture synthesis)",
    "phase4: Phase 4 tests (development planning)",
    "phase5: Phase 5 tests (code generation)",
    "e2e: End-to-end tests (require LLM/ChromaDB)",
    "slow: Slow tests (>10 seconds)",
]
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.ruff]
target-version = "py312"
line-length = 120
src = ["src", "tests", "ui/backend"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "S", "RUF"]
ignore = [
    "B008",    # Function call in argument defaults (FastAPI Depends)
    "B904",    # raise ... from err in except clauses (too pervasive to fix now)
    "E402",    # Module-level import not at top (intentional: filter warnings before imports)
    "E722",    # Bare except (used in JSON parsing fallbacks)
    "RUF012",  # Mutable class variables (CrewAI pattern)
    "RUF013",  # Implicit Optional (would touch hundreds of signatures)
    "S101",    # Use of assert (OK in tests)
    "S110",    # try-except-pass (intentional in metric logging)
    "S104",    # Binding to all interfaces (intentional for Docker)
    "S112",    # try-except-continue (intentional in batch processing)
    "S314",    # XML parsing (internal files only)
    "S324",    # Insecure hash (not for security purposes)
    "S603",    # subprocess call (internal commands only)
    "S607",    # Partial executable path (git, python — trusted)
    "E501",    # Line too long (handled by formatter where possible)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "B017"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

[tool.crewai]
type = "crew"
